backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
chromadb==0.5.5
sentence-transformers==3.0.1
anthropic==0.34.1
requests==2.31.0
langchain==0.2.16
langchain-community==0.2.16
python-dotenv==1.0.1
pydantic==2.9.1
